+++ src/jest.ts
@@ -5,16 +5,21 @@ import { JestConfigWithTsJest } from "ts-jest";
  */
 export const jest: JestConfigWithTsJest = {
   collectCoverage: true,
-  extensionsToTreatAsEsm: [".ts"],
-  // Jest doesn't resolve import/export statements ending with a .js into .jsx, .ts, or .tsx
-  // and leaves them as is resulting in a "Cannot find module" error. By stripping the .js
-  // off the end, it becomes a "best match" and we just get lucky.
-  // https://github.com/facebook/jest/issues/9430#issuecomment-836760438
+  testEnvironmentOptions: {
+    url: "http://localhost/",
+  },
+
+  // Copied from https://kulshekhar.github.io/ts-jest/docs/guides/esm-support
+  preset: "ts-jest/presets/default-esm",
   moduleNameMapper: {
     "^(\\.{1,2}/.*)\\.js$": "$1",
   },
-  preset: "ts-jest/presets/default-esm",
-  testEnvironmentOptions: {
-    url: "http://localhost/",
+  transform: {
+    "^.+\\.m?[t]sx?$": [
+      "ts-jest",
+      {
+        useESM: true,
+      },
+    ],
   },
 };