+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed jest test config to correctly apply ts-jest's ESM config
+
 ## [3.0.0-beta.8] - 2022-12-30
 
 - Fixed release-it jsdoc comments